# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Agora API
  version: 0.1.0
  description: >-
    Authored public contract for the implemented Phase 1 API. Runtime OpenAPI routes
    remain disabled so this reviewed document is the only public API description.
servers:
  - url: http://127.0.0.1:8000
    description: Local development
...
paths:
...
  /api/v1/formalizations:
    post:
      operationId: createFormalization
      tags: [formalizations]
      summary: Create an immutable candidate formalization
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FormalizationCreate" }
      responses:
        "201": { $ref: "#/components/responses/FormalizationWritten" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/formalizations/{formalization_id}:
    get:
      operationId: getFormalization
      tags: [formalizations]
      summary: Read the current or selected immutable revision
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FormalizationId"
        - name: revision
          in: query
          schema: { type: integer, minimum: 1 }
      responses:
        "200": { $ref: "#/components/responses/FormalizationRead" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/formalizations/{formalization_id}/revisions:
    post:
      operationId: reviseFormalization
      tags: [formalizations]
      summary: Append a replacement content revision
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FormalizationId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FormalizationCreate" }
      responses:
        "201": { $ref: "#/components/responses/FormalizationWritten" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/formalizations/{formalization_id}/validations:
    post:
      operationId: validateFormalization
      tags: [formalizations]
      summary: Record deterministic structural validation
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FormalizationId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: false }
      responses:
        "201": { $ref: "#/components/responses/FormalizationWritten" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/formalizations/{formalization_id}/confirmations:
    post:
      operationId: confirmFormalization
      tags: [formalizations]
      summary: Record human confirmation after successful validation
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FormalizationId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FormalizationDecisionCreate" }
      responses:
        "201": { $ref: "#/components/responses/FormalizationWritten" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/formalizations/{formalization_id}/rejections:
    post:
      operationId: rejectFormalization
      tags: [formalizations]
      summary: Record an explicit human rejection
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FormalizationId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FormalizationDecisionCreate" }
      responses:
        "201": { $ref: "#/components/responses/FormalizationWritten" }
        "409": { $ref: "#/components/responses/Conflict" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  headers:
    ETag:
      description: Current integer entity version.
      schema:
        type: string
        pattern: "^[1-9][0-9]*$"
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 255
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: integer
        minimum: 1
    FormalizationId:
      name: formalization_id
      in: path
      required: true
      schema: { $ref: "#/components/schemas/FormalizationId" }
  responses:
    FormalizationRead:
      description: The authoritative immutable formalization revision and derived status.
      headers:
        ETag: { $ref: "#/components/headers/ETag" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/FormalizationResponse" }
    FormalizationWritten:
      description: The formalization revision or lifecycle fact committed atomically.
      headers:
        ETag: { $ref: "#/components/headers/ETag" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/FormalizationResponse" }
...
    ValidationFailed:
      description: The strict request or a domain invariant failed validation.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    NotFound:
      description: The resource is absent or hidden by tenant isolation.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Conflict:
      description: The entity version or idempotency request identity conflicts.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
  schemas:
...
    ArtifactId:
      type: string
      pattern: "^art_[0-9a-f]{32}$"
    FormalizationId:
      type: string
      pattern: "^frm_[0-9a-f]{32}$"
    ValidationStatus:
      type: string
      enum: [CANDIDATE, VALIDATED, REJECTED]
    SymbolDeclaration:
      type: object
      additionalProperties: false
      required: [name, sort, meaning, unit]
      properties:
        name: { type: string, pattern: "^[A-Za-z][A-Za-z0-9_]{0,63}$" }
        sort: { type: string, enum: [BOOLEAN, INTEGER, REAL] }
        meaning: { type: string, minLength: 1 }
        unit: { type: [string, "null"] }
    FormalAST:
      description: Closed typed AST. Unknown nodes and operators are rejected.
      oneOf:
        - $ref: "#/components/schemas/FormalBoolean"
        - $ref: "#/components/schemas/FormalInteger"
        - $ref: "#/components/schemas/FormalReal"
        - $ref: "#/components/schemas/FormalSymbol"
        - $ref: "#/components/schemas/FormalOperation"
      discriminator:
        propertyName: kind
    FormalBoolean:
      type: object
      additionalProperties: false
      required: [kind, value]
      properties:
        kind: { type: string, const: boolean }
        value: { type: boolean }
    FormalInteger:
      type: object
      additionalProperties: false
      required: [kind, value]
      properties:
        kind: { type: string, const: integer }
        value: { type: string, pattern: "^(?:0|-?[1-9][0-9]*)$" }
    FormalReal:
      type: object
      additionalProperties: false
      required: [kind, value]
      properties:
        kind: { type: string, const: real }
        value: { type: string, pattern: "^(?:0|-?[1-9][0-9]*)(?:\\.[0-9]*[1-9])?$" }
    FormalSymbol:
      type: object
      additionalProperties: false
      required: [kind, name]
      properties:
        kind: { type: string, const: symbol }
        name: { type: string, pattern: "^[A-Za-z][A-Za-z0-9_]{0,63}$" }
    FormalOperation:
      type: object
      additionalProperties: false
      required: [kind, operator, arguments]
      properties:
        kind: { type: string, const: operation }
        operator:
          type: string
          enum: [NOT, AND, OR, NEG, ADD, SUB, MUL, DIV, EQ, NE, LT, LE, GT, GE]
        arguments:
          type: array
          items: { $ref: "#/components/schemas/FormalAST" }
    FormalizationCreate:
      type: object
      additionalProperties: false
      required: [source_artifact_id, ast, symbols, canonical_rendering, premise_artifact_ids, limitations, fidelity_notes]
      properties:
        source_artifact_id: { $ref: "#/components/schemas/ArtifactId" }
        ast: { $ref: "#/components/schemas/FormalAST" }
        symbols:
          type: array
          items: { $ref: "#/components/schemas/SymbolDeclaration" }
        canonical_rendering: { type: string, minLength: 1 }
        premise_artifact_ids:
          type: array
          uniqueItems: true
          items: { $ref: "#/components/schemas/ArtifactId" }
        limitations:
          type: array
          minItems: 1
          items: { type: string, minLength: 1 }
        fidelity_notes: { type: string, minLength: 1 }
    FormalizationDecisionCreate:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason: { type: string, minLength: 1 }
    FormalizationResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [id, revision_id, revision, ast, ast_hash, validation_status, enforceable]
          properties:
            id: { $ref: "#/components/schemas/FormalizationId" }
            revision_id: { $ref: "#/components/schemas/FormalizationId" }
            revision: { type: integer, minimum: 1 }
            ast: { $ref: "#/components/schemas/FormalAST" }
            ast_hash: { type: string, pattern: "^sha256:[0-9a-f]{64}$" }
            validation_status: { $ref: "#/components/schemas/ValidationStatus" }
            enforceable: { type: boolean }
          additionalProperties: true
        meta: { type: object, additionalProperties: true }
...
    ValidationIssue:
      type: object
      additionalProperties: false
      required: [pointer, message, type]
      properties:
        pointer:
          type: string
        message:
          type: string
        type:
          type: string
    ErrorCode:
      type: string
      enum:
        - VALIDATION_FAILED
        - UNKNOWN_FIELD
        - AUTH_REQUIRED
        - FORBIDDEN
        - COMMIT_FORBIDDEN
        - VERSION_CONFLICT
        - IDEMPOTENCY_REPLAY
        - NOT_FOUND
        - PROVENANCE_MISSING
        - BUDGET_EXCEEDED
        - RATE_LIMITED
        - PORT_UNAVAILABLE
        - RAG_FAILED
        - WORKFLOW_UNREACHABLE
        - INTERNAL
    ProblemDetails:
      type: object
      required: [type, title, status, code, detail, retryable]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        code:
          $ref: "#/components/schemas/ErrorCode"
        detail:
          type: string
        instance:
          type: [string, "null"]
        trace_id:
          type: [string, "null"]
        retryable:
          type: boolean
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ValidationIssue"
        current_version:
          type: integer
        retry_after_s:
          type: number
        degraded:
          type: string